Move persistence connector setup store from internal to connectorsetupstore package.
